+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for Blade component namespaces (i.e. `<x-module::component.name />`)
+
 ## [1.7.0] - 2022-02-11
 
 ### Added